backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
email-validator==2.1.0.post1
fastapi-mail==1.4.1